feat(conv): connected UI to provider
- can start and stop a normal conversation with no context
- left the /conversation/example page free from provider which can be
  used to finetune the UI responsivness

closes AIS-59, AIS-60
