+++ src/app/(account)/conversation/page.tsx
@@ -1,98 +1,37 @@
 "use client";
 
-import { useEffect, useState } from "react";
-
 import { Button } from "@/components/ui/button";
-import { Orb } from "@/features/conversation/components/orb";
+import { Orb, useConversation } from "@/features/conversation";
 
 // Example usage component
 const ConversationPage = () => {
-	const [isActive, setIsActive] = useState(false);
-	const [isTalking, setIsTalking] = useState(false);
-	const [talkingIntensity, setTalkingIntensity] = useState(0);
-
-	useEffect(() => {
-		console.log("isTalking changed:", isTalking); // Debug log
-		let startTime = Date.now() / 1000;
-		let animationFrameId: number;
-
-		const updateIntensity = () => {
-			const currentTime = Date.now() / 1000;
-			const elapsedTime = currentTime - startTime;
-
-			// Scale the elapsed time to match our time array range
-			const scaledTime = elapsedTime + time[0]; // Add the offset to start from the first time value
-
-			// Find the two closest time points for interpolation
-			const nextIndex = time.findIndex((t) => t > scaledTime);
-
-			if (nextIndex === -1) {
-				console.log("Reached end of sequence"); // Debug log
-				setTalkingIntensity(intensity[intensity.length - 1]);
-				setIsTalking(false);
-				return;
-			}
-
-			if (nextIndex === 0) {
-				// Before first time point, use first intensity
-				setTalkingIntensity(intensity[0]);
-			} else {
-				// Interpolate between the two closest points
-				const prevTime = time[nextIndex - 1];
-				const nextTime = time[nextIndex];
-				const prevIntensity = intensity[nextIndex - 1];
-				const nextIntensity = intensity[nextIndex];
-
-				// Calculate interpolation factor (0 to 1)
-				const t = (scaledTime - prevTime) / (nextTime - prevTime);
-
-				// Linear interpolation between intensities
-				const interpolatedIntensity =
-					prevIntensity + t * (nextIntensity - prevIntensity);
-				setTalkingIntensity(interpolatedIntensity);
-			}
-
-			animationFrameId = requestAnimationFrame(updateIntensity);
-		};
-
-		if (isTalking) {
-			startTime = Date.now() / 1000;
-			updateIntensity();
-		} else {
-			setTalkingIntensity(0);
-		}
-
-		return () => {
-			if (animationFrameId) {
-				cancelAnimationFrame(animationFrameId);
-			}
-		};
-	}, [isTalking]);
+	const {
+		isSessionActive,
+		isAgentSpeaking,
+		audioIntensity,
+		startSession,
+		stopSession,
+	} = useConversation();
 
 	return (
 		<div className="flex min-h-screen w-full flex-col justify-center">
 			<Orb
-				isActive={isActive}
-				isTalking={isTalking}
-				talkingIntensity={talkingIntensity}
+				isActive={isSessionActive}
+				isTalking={isAgentSpeaking}
+				talkingIntensity={audioIntensity}
 			/>
-			<div className="mt-4 flex gap-4">
+			<div className="mt-4 flex justify-center gap-4">
 				<Button
 					onClick={() => {
-						setIsActive(!isActive);
+						if (isSessionActive) {
+							stopSession();
+						} else {
+							startSession();
+						}
 					}}
 					className="border border-white bg-black text-white hover:bg-gray-800"
 				>
-					{isActive ? "Deactivate" : "Activate"}
-				</Button>
-				<Button
-					onClick={() => {
-						setIsTalking(!isTalking);
-					}}
-					disabled={!isActive}
-					className="border border-white bg-black text-white hover:bg-gray-800 disabled:opacity-50"
-				>
-					{isTalking ? "Stop Talking" : "Simulate Talking"}
+					{isSessionActive ? "Deactivate" : "Activate"}
 				</Button>
 			</div>
 		</div>
@@ -100,78 +39,3 @@ const ConversationPage = () => {
 };
 
 export default ConversationPage;
-
-const intensity = [
-	0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5.546875, 27.9921875, 45.140625, 54.6171875,
-	60.9296875, 62.890625, 63.90625, 64.046875, 64.4921875, 65.890625, 67.40625,
-	69.8046875, 71.09375, 68.4921875, 65.4765625, 62.59375, 60.9921875,
-	61.4921875, 64.03125, 66.5703125, 67.84375, 68.1953125, 67.2421875, 66.671875,
-	64.7578125, 62.9453125, 61.421875, 59.2890625, 56.9765625, 54.3828125,
-	51.8828125, 49.1875, 46.6796875, 43.953125, 40.640625, 37.1484375, 33.8125,
-	30.5546875, 27.3359375, 24.234375, 21.296875, 18.4609375, 15.765625,
-	13.2265625, 10.921875, 8.875, 7.1640625, 6.046875, 5.21875, 4.578125,
-	3.8984375, 3.28125, 2.8046875, 37.875, 48.6875, 54.265625, 59.3671875, 63.75,
-	64.609375, 64.875, 62.4921875, 59.75, 56.6953125, 53.5, 50.1171875, 46.890625,
-	61.015625, 66.40625, 67.484375, 66.84375, 65.8046875, 63.5, 61.5, 60.3125, 59,
-	58.71875, 59.53125, 56.34375, 53.359375, 50.2734375, 47.25, 43.921875,
-	41.1640625, 46.0234375, 48.265625, 51.703125, 53.6953125, 53.53125,
-	53.1640625, 52.078125, 52.640625, 53.390625, 56.7109375, 61.21875, 63.15625,
-	64, 63.6953125, 63.671875, 62.9375, 62.15625, 59.3046875, 55.78125,
-	52.2578125, 48.71875, 45.1953125, 41.7265625, 38.375, 35.15625, 32.359375,
-	32.5234375, 37.203125, 44.0703125, 50.5625, 54.1875, 57.8359375, 61.328125,
-	64.15625, 64.2265625, 62.5078125, 60.640625, 60.8203125, 75.453125, 79.90625,
-	81.140625, 78.703125, 75.6015625, 72.46875, 70.140625, 69.40625, 68.828125,
-	66.8515625, 64.3984375, 61.9296875, 61.4375, 62.28125, 63.8203125, 66.4140625,
-	66.8984375, 68.4140625, 69.6328125, 70.9140625, 73.5390625, 73.6953125,
-	79.1953125, 83.4765625, 84.1796875, 85.171875, 85.8203125, 82.4453125,
-	78.78125, 75.125, 73.28125, 73.6171875, 73.078125, 71.625, 68.75, 65.546875,
-	62.1015625, 58.640625, 55.09375, 51.5625, 48.0859375, 44.5703125, 41.0234375,
-	37.5546875, 34.09375, 30.59375, 27.2890625, 23.96875, 20.7109375, 17.546875,
-	14.6640625, 12.0234375, 9.625, 7.4453125, 5.7734375, 4.59375, 4.8046875,
-	23.9609375, 42.78125, 52.7265625, 58.515625, 61.6484375, 61.9375, 62.203125,
-	58.8046875, 55.34375, 53.46875, 56.4375, 58.9921875, 60.6640625, 61.7578125,
-	64.15625, 64.4765625, 63.875, 63.875, 61.859375, 60.2734375, 59.1875,
-	59.265625, 59.453125, 63.1015625, 65.6328125, 66.78125, 67.2578125, 66.671875,
-	66.6796875, 65.859375, 64.3671875, 63.078125, 61.7890625, 59.5859375,
-	56.8359375, 54.171875, 50.9375, 47.6171875, 44.140625, 41.0859375, 37.6953125,
-	34.3125, 30.859375, 27.6953125, 24.6484375, 21.796875, 19.0859375, 16.453125,
-	13.953125, 11.6953125, 9.6328125, 7.6953125, 6.125, 5.078125, 4.296875,
-	3.71875, 3.3671875, 6.2421875, 11.6875, 17.3984375, 34.375, 47.40625,
-	59.0546875, 65.9296875, 67.4921875, 64.375, 60.828125, 62.90625, 69.984375,
-	76.6015625, 80.1015625, 79.9609375, 79.3828125, 80.234375, 79.796875,
-	79.265625, 78.078125, 76.8828125, 75.4453125, 73.890625, 73.703125,
-	71.8828125, 68.4296875, 64.90625, 61.3125, 57.71875, 54.515625, 52.6875,
-	49.6015625, 46.328125, 42.8671875, 39.34375, 35.8046875, 32.3828125,
-	29.0703125, 25.890625, 22.609375, 19.4140625, 16.4453125, 13.640625,
-	10.828125, 8.25, 6.28125, 4.96875, 4.1328125, 3.5234375, 3.0078125, 2.515625,
-	2.0234375, 1.5390625, 1.1015625, 0.6640625, 0.265625, 0.0390625, 0, 0, 0, 0,
-	0, 0, 0, 0, 0, 0,
-];
-
-const time = [
-	3.01, 3.03, 3.05, 3.06, 3.08, 3.1, 3.11, 3.13, 3.15, 3.16, 3.18, 3.2, 3.21,
-	3.23, 3.25, 3.26, 3.28, 3.3, 3.31, 3.33, 3.35, 3.36, 3.38, 3.4, 3.41, 3.43,
-	3.45, 3.46, 3.48, 3.5, 3.51, 3.53, 3.55, 3.56, 3.58, 3.6, 3.62, 3.63, 3.65,
-	3.66, 3.68, 3.7, 3.71, 3.73, 3.75, 3.76, 3.78, 3.8, 3.81, 3.83, 3.85, 3.86,
-	3.88, 3.9, 3.91, 3.93, 3.95, 3.96, 3.98, 4.0, 4.01, 4.03, 4.05, 4.06, 4.08,
-	4.1, 4.11, 4.13, 4.15, 4.16, 4.18, 4.2, 4.21, 4.23, 4.25, 4.26, 4.28, 4.3,
-	4.31, 4.33, 4.35, 4.36, 4.38, 4.4, 4.41, 4.43, 4.45, 4.46, 4.48, 4.5, 4.51,
-	4.53, 4.55, 4.56, 4.58, 4.6, 4.61, 4.63, 4.65, 4.66, 4.68, 4.7, 4.71, 4.73,
-	4.75, 4.76, 4.78, 4.8, 4.81, 4.83, 4.85, 4.86, 4.88, 4.9, 4.91, 4.93, 4.95,
-	4.96, 4.98, 5.0, 5.01, 5.03, 5.05, 5.06, 5.08, 5.1, 5.11, 5.13, 5.15, 5.16,
-	5.18, 5.2, 5.21, 5.23, 5.25, 5.26, 5.28, 5.3, 5.31, 5.33, 5.35, 5.36, 5.38,
-	5.4, 5.41, 5.43, 5.45, 5.46, 5.48, 5.5, 5.51, 5.53, 5.55, 5.56, 5.58, 5.6,
-	5.61, 5.63, 5.65, 5.66, 5.68, 5.7, 5.71, 5.73, 5.75, 5.76, 5.78, 5.8, 5.81,
-	5.83, 5.85, 5.86, 5.88, 5.9, 5.91, 5.93, 5.95, 5.96, 5.98, 6.0, 6.01, 6.03,
-	6.05, 6.06, 6.08, 6.1, 6.11, 6.13, 6.15, 6.16, 6.18, 6.2, 6.21, 6.23, 6.25,
-	6.26, 6.28, 6.29, 6.31, 6.33, 6.35, 6.36, 6.38, 6.4, 6.41, 6.43, 6.45, 6.46,
-	6.48, 6.5, 6.51, 6.53, 6.55, 6.56, 6.58, 6.6, 6.61, 6.63, 6.65, 6.66, 6.68,
-	6.7, 6.71, 6.73, 6.75, 6.76, 6.78, 6.8, 6.81, 6.83, 6.85, 6.86, 6.88, 6.9,
-	6.91, 6.93, 6.95, 6.96, 6.98, 7.0, 7.01, 7.03, 7.05, 7.06, 7.08, 7.1, 7.11,
-	7.13, 7.15, 7.16, 7.18, 7.2, 7.21, 7.23, 7.25, 7.26, 7.28, 7.3, 7.31, 7.33,
-	7.35, 7.36, 7.38, 7.4, 7.41, 7.43, 7.45, 7.46, 7.48, 7.5, 7.51, 7.53, 7.55,
-	7.56, 7.58, 7.6, 7.61, 7.63, 7.65, 7.66, 7.68, 7.7, 7.71, 7.73, 7.75, 7.76,
-	7.78, 7.8, 7.81, 7.83, 7.85, 7.86, 7.88, 7.9, 7.92, 7.93, 7.95, 7.96, 7.98,
-	8.0, 8.01, 8.03, 8.05, 8.06, 8.08, 8.1, 8.11, 8.13, 8.15, 8.16, 8.18, 8.2,
-	8.21,
-];
