+++ src/features/conversation/components/index.ts
@@ -1,3 +1,2 @@
-export { default as ConversationControls } from "./conversation-controls";
-export { default as ConversationOrb } from "./conversation-orb";
 export { default as DustCloud } from "./dustcloud";
+export { default as Orb } from "./orb";
